fix(java): delete tag before release
